requirements.txt
discord.py==2.1.0
py-cord==2.3.2
